requirements.txt
scrapy
psycopg2-binary
dbt-postgres


